% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofeatureR-package.R
\docType{package}
\name{cofeatureR}
\alias{cofeatureR}
\alias{cofeatureR-package}
\title{cofeatureR: Generate Cofeature Matrices}
\description{
Generate cofeature (feature by sample) matrices. The package 
utilizes ggplot2::geom_tile to generate the matrix allowing for easy
customization of additions from the base matrix.
}
